+++ ReizenNaarGentState.java
@@ -1,3 +1,5 @@
+//TODO original deprecated
+
 import java.util.Map;
 
 public class ReizenNaarGentState extends State {
@@ -10,10 +12,10 @@ public class ReizenNaarGentState extends State {
 
     @Override
     public State verwerkKeuze(int keuze, StudentStats stats) {
-        Map<String, Integer> map =stats.afvalProductie;
+        Map<String, Double> map =stats.afvalProductie;
         if (keuze == 1) {
             stats.co2Uitstoot += 30;
-            map.put("Industrieel afval", map.getOrDefault("Industrieel afval", 0) + 1);
+            map.put("Industrieel afval", map.getOrDefault("Industrieel afval", 0.0) + 1);
         } else if (keuze == 2) {
             stats.co2Uitstoot += 5;
         }
